Refactor add-new-project.js and delete-edit-pop-up.js code
Rewrite the code for add-new-project.js to get it's project name
from projectArray items instead of getting it directly from input
field as it was before.

Also rewrite the code for delete-edit-pop-up.js so it can delete
project from projectArray while it's being removed from DOM.

Also use "bind" to bind the value of element eventListener is
added to the value of "this" from the function used.