Refactor all-tasks.js to be able to toggle states
After I created the all-tasks.js module with its functions i was
unable to toggle states from the all tasks tab.

This was beacuse currentProjectArray which is used by the Project
tab was not useable in the all-tasks.js module so i had to create
a seperate array for the module.

In the toggle-states.js I added an if statement to look for index
from allTasksArray whenever the currentProjectArray returns -1
which indicates that the index is not found there so it should
look for it from allTaskArray in all-tasks.js